+++ lib/vump/modules/composer_json.rb
@@ -1,0 +1,9 @@
+require 'vump/modules/package_json'
+
+module Vump
+  class ComposerJson < PackageJson
+    def filename
+      'composer.json'
+    end
+  end
+end