+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
  - Vump badge
+ - Bower module
+ - Component module
+ - Composer module
  - Git ignore module support
 
 ## [2.0.0] - 2019-03-14